Deprecate Brew Evolution.
This repository hasn't really served it's purpose as not all proposals
have been implemented or sufficiently prioritised. There's been
insufficient community interest in this process and it's much heavier
than discussion on an issue. Additionally, it can provide a "set in
stone" feeling where design has been agreed but there were unforeseen
problems with the implementation.